--- jsconfig.json
@@ -1,6 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2020"
-  },
-  "exclude": ["node_modules"]
-}
