+++ scripts/background.ts
@@ -1,4 +1,4 @@
-'use strict';
+import WebDAVFS from './webdav_fs';
 
 (() => {
   if (!chrome.fileSystemProvider) {
@@ -13,13 +13,15 @@
       url: 'window.html',
       type: 'popup',
     });
+    if (!window.id) return;
     await browser.windows.update(window.id, {
       width: 400, height: 300,
     });
   }
 
-  async function mount(request) {
-    const mounted = await fs.isMounted(request.url, request.username);
+  async function mount(request: any) {
+    const { url, username } = request;
+    const mounted = await fs.isMounted(url, username);
     if (mounted) throw new Error('Already mounted');
     await fs.mount(request);
   }
@@ -28,9 +30,9 @@
     await browser.storage.local.set({ version: 'v1' });
   });
 
-  browser.fileSystemProvider.onMountRequested.addListener(openWindow);
-
   browser.runtime.onMessage.addListener(mount);
+
+  chrome.fileSystemProvider.onMountRequested.addListener(openWindow);
 
   // fs.resumeMounts(); // for debug
 })();
